+++ package.json
@@ -13,7 +13,8 @@
     "next": "9.1.2",
     "react": "16.11.0",
     "react-dom": "16.11.0",
-    "react-redux": "^7.1.3"
+    "react-redux": "^7.1.3",
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
     "@types/node": "^12.12.6",
